requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
Markdown==3.1.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
